SquirrelPAK 1201 - Tech Squirrels v0.2.0-1.20.1
SquirrelPAK 1201 Release 0.2.0
File Version 0.2.0-1.20.1

Required Forge Version: 47.3.7

Beta 2

Summary:
Updated mods that had updates and added
Alchemistry! Since it is open source we can
include it here. Check the credits and
licenses files for details

Also fixed the issue with BCC and added it
back to the PAK. Was an issue with the mod
how it was added to Modrinth the newest ver
of the mod is not the default.